+++ analytics.js
@@ -393,121 +393,13 @@ function parse(str) {
 }
 
 });
-require.register("component-to-function/index.js", function(exports, require, module){
-
-/**
- * Expose `toFunction()`.
- */
-
-module.exports = toFunction;
-
-/**
- * Convert `obj` to a `Function`.
- *
- * @param {Mixed} obj
- * @return {Function}
- * @api private
- */
-
-function toFunction(obj) {
-  switch ({}.toString.call(obj)) {
-    case '[object Object]':
-      return objectToFunction(obj);
-    case '[object Function]':
-      return obj;
-    case '[object String]':
-      return stringToFunction(obj);
-    case '[object RegExp]':
-      return regexpToFunction(obj);
-    default:
-      return defaultToFunction(obj);
-  }
-}
-
-/**
- * Default to strict equality.
- *
- * @param {Mixed} val
- * @return {Function}
- * @api private
- */
-
-function defaultToFunction(val) {
-  return function(obj){
-    return val === obj;
-  }
-}
-
-/**
- * Convert `re` to a function.
- *
- * @param {RegExp} re
- * @return {Function}
- * @api private
- */
-
-function regexpToFunction(re) {
-  return function(obj){
-    return re.test(obj);
-  }
-}
-
-/**
- * Convert property `str` to a function.
- *
- * @param {String} str
- * @return {Function}
- * @api private
- */
-
-function stringToFunction(str) {
-  // immediate such as "> 20"
-  if (/^ *\W+/.test(str)) return new Function('_', 'return _ ' + str);
-
-  // properties such as "name.first" or "age > 18"
-  return new Function('_', 'return _.' + str);
-}
-
-/**
- * Convert `object` to a function.
- *
- * @param {Object} object
- * @return {Function}
- * @api private
- */
-
-function objectToFunction(obj) {
-  var match = {}
-  for (var key in obj) {
-    match[key] = typeof obj[key] === 'string'
-      ? defaultToFunction(obj[key])
-      : toFunction(obj[key])
-  }
-  return function(val){
-    if (typeof val !== 'object') return false;
-    for (var key in match) {
-      if (!(key in val)) return false;
-      if (!match[key](val[key])) return false;
-    }
-    return true;
-  }
-}
-
-});
 require.register("component-each/index.js", function(exports, require, module){
 
 /**
  * Module dependencies.
  */
 
-var toFunction = require('to-function');
-var type;
-
-try {
-  type = require('type-component');
-} catch (e) {
-  type = require('type');
-}
+var type = require('type');
 
 /**
  * HOP reference.
@@ -524,7 +416,6 @@ var has = Object.prototype.hasOwnProperty;
  */
 
 module.exports = function(obj, fn){
-  fn = toFunction(fn);
   switch (type(obj)) {
     case 'array':
       return array(obj, fn);
@@ -579,7 +470,6 @@ function array(obj, fn) {
     fn(obj[i], i);
   }
 }
-
 });
 require.register("component-event/index.js", function(exports, require, module){
 
@@ -838,15 +728,15 @@ exports.parse = function(url){
   a.href = url;
   return {
     href: a.href,
-    host: a.host || location.host,
-    port: ('0' === a.port || '' === a.port) ? location.port : a.port,
+    host: a.host,
+    port: a.port,
     hash: a.hash,
-    hostname: a.hostname || location.hostname,
-    pathname: a.pathname.charAt(0) != '/' ? '/' + a.pathname : a.pathname,
-    protocol: !a.protocol || ':' == a.protocol ? location.protocol : a.protocol,
+    hostname: a.hostname,
+    pathname: a.pathname,
+    protocol: a.protocol,
     search: a.search,
     query: a.search.slice(1)
-  };
+  }
 };
 
 /**
@@ -858,7 +748,9 @@ exports.parse = function(url){
  */
 
 exports.isAbsolute = function(url){
-  return 0 == url.indexOf('//') || !!~url.indexOf('://');
+  if (0 == url.indexOf('//')) return true;
+  if (~url.indexOf('://')) return true;
+  return false;
 };
 
 /**
@@ -870,7 +762,7 @@ exports.isAbsolute = function(url){
  */
 
 exports.isRelative = function(url){
-  return !exports.isAbsolute(url);
+  return ! exports.isAbsolute(url);
 };
 
 /**
@@ -883,9 +775,9 @@ exports.isRelative = function(url){
 
 exports.isCrossDomain = function(url){
   url = exports.parse(url);
-  return url.hostname !== location.hostname
-    || url.port !== location.port
-    || url.protocol !== location.protocol;
+  return url.hostname != location.hostname
+    || url.port != location.port
+    || url.protocol != location.protocol;
 };
 });
 require.register("ianstormtaylor-callback/index.js", function(exports, require, module){
@@ -1318,6 +1210,34 @@ module.exports = function canonical () {
     if ('canonical' == tag.getAttribute('rel')) return tag.getAttribute('href');
   }
 };
+});
+require.register("segmentio-convert-dates/index.js", function(exports, require, module){
+
+var is = require('is');
+
+
+/**
+ * Expose `convertDates`.
+ */
+
+module.exports = convertDates;
+
+
+/**
+ * Recursively convert an `obj`'s dates to new values.
+ *
+ * @param {Object} obj
+ * @param {Function} convert
+ * @return {Object}
+ */
+
+function convertDates (obj, convert) {
+  for (var key in obj) {
+    var val = obj[key];
+    if (is.date(val)) obj[key] = convert(val);
+    if (is.object(val)) convertDates(val, convert);
+  }
+}
 });
 require.register("segmentio-extend/index.js", function(exports, require, module){
 
@@ -4532,94 +4452,126 @@ Inspectlet.prototype.initialize = function (options, ready) {
 };
 });
 require.register("analytics/lib/providers/intercom.js", function(exports, require, module){
-// http://docs.intercom.io/
-// http://docs.intercom.io/#IntercomJS
 
-var Provider = require('../provider')
-  , extend   = require('extend')
-  , load     = require('load-script')
-  , isEmail  = require('is-email');
+var alias = require('alias')
+  , convertDates = require('convert-dates')
+  , integration = require('../integration')
+  , each = require('each')
+  , is = require('is')
+  , isEmail = require('is-email')
+  , load = require('load-script');
 
 
-module.exports = Provider.extend({
+/**
+ * Expose `Intercom` integration.
+ *
+ * http://docs.intercom.io/
+ * http://docs.intercom.io/#IntercomJS
+ */
 
-  name : 'Intercom',
+var Intercom = module.exports = integration('Intercom');
 
-  // Whether Intercom has already been booted or not. Intercom becomes booted
-  // after Intercom('boot', ...) has been called on the first identify.
-  booted : false,
 
-  key : 'appId',
+/**
+ * Required key.
+ */
 
-  defaults : {
-    // Intercom's required key.
-    appId : null,
-    // An optional setting to display the Intercom inbox widget.
-    activator : null,
-    // Whether to show the count of messages for the inbox widget.
-    counter : true
-  },
+Intercom.prototype.key = 'appId';
 
-  initialize : function (options, ready) {
-    load('https://static.intercomcdn.com/intercom.v1.js', ready);
-  },
 
-  identify : function (userId, traits, options) {
-    // Don't do anything if we just have traits the first time.
-    if (!this.booted && !userId) return;
+/**
+ * Default options.
+ */
 
-    // Intercom specific settings. BACKWARDS COMPATIBILITY: we need to check for
-    // the lowercase variant as well.
-    options || (options = {});
-    var Intercom = options.Intercom || options.intercom || {};
-    traits.increments = Intercom.increments;
-    traits.user_hash = Intercom.userHash || Intercom.user_hash;
+Intercom.prototype.defaults = {
+  // an optional css selector to use for the intercom inbox widget button
+  activator: '',
+  // your intercom app id (required)
+  appId: '',
+  // whether to show the count of messages on the intercom inbox widget
+  counter: true,
+  // whether or not to show the intercom inbox widget
+  inbox: false
+};
 
-    // They need `created_at` as a Unix timestamp (seconds).
-    if (traits.created) {
-      traits.created_at = Math.floor(traits.created/1000);
-      delete traits.created;
-    }
 
-    // Convert a `company`'s `created` date.
-    if (traits.company && traits.company.created) {
-      traits.company.created_at = Math.floor(traits.company.created/1000);
-      delete traits.company.created;
-    }
+/**
+ * Initialize.
+ *
+ * @param {Object} options
+ * @param {Function} ready
+ */
 
-    // Optionally add the inbox widget.
-    if (this.options.activator) {
-      traits.widget = {
-        activator   : this.options.activator,
-        use_counter : this.options.counter
-      };
-    }
+Intercom.prototype.initialize = function (options, ready) {
+  load('https://static.intercomcdn.com/intercom.v1.js', ready);
+};
 
-    // If this is the first time we've identified, `boot` instead of `update`
-    // and add our one-time boot settings.
-    if (this.booted) {
-      window.Intercom('update', traits);
-    } else {
-      extend(traits, {
-        app_id  : this.options.appId,
-        user_id : userId
-      });
-      window.Intercom('boot', traits);
-    }
 
-    // Set the booted state, so that we know to call 'update' next time.
-    this.booted = true;
-  },
+/**
+ * Identify.
+ *
+ * @param {String} id (optional)
+ * @param {Object} traits (optional)
+ * @param {Object} options (optional)
+ */
 
-  // Intercom doesn't have a separate `group` method, but they take a
-  // `companies` trait for the user.
-  group : function (groupId, properties, options) {
-    properties.id = groupId;
-    window.Intercom('update', { company : properties });
+Intercom.prototype.identify = function (id, traits, options) {
+  var method = this._id !== id ? 'boot': 'update';
+  this._id = id; // cache for next time
+
+  // required options
+  traits.app_id = this.options.appId;
+  if (id) traits.user_id = id;
+  if (isEmail(id) && !traits.email) traits.email = id;
+  if (!traits.user_id && !traits.email) return;
+
+  // handle dates
+  convertDates(traits, convertDate);
+  alias(traits, { created: 'created_at'});
+  if (traits.company) alias(traits.company, { created: 'created_at' });
+
+  // handle options
+  options || (options = {});
+  var Intercom = options.Intercom || options.intercom || {};
+  if (Intercom.increments) traits.increments = Intercom.increments;
+  if (Intercom.userHash) traits.user_hash = Intercom.userHash;
+  if (Intercom.user_hash) traits.user_hash = Intercom.user_hash;
+    // TODO: make this activator's default and run a migration
+  if (this.options.inbox || this.options.activator) {
+    traits.widget = {
+      activator: this.options.activator || '#Intercom',
+      use_counter: this.options.counter
+    };
   }
 
-});
+  window.Intercom(method, traits);
+};
 
+
+/**
+ * Group.
+ *
+ * @param {String} id
+ * @param {Object} properties (optional)
+ * @param {Object} options (optional)
+ */
+
+Intercom.prototype.group = function (id, properties, options) {
+  properties.id = id;
+  window.Intercom('update', { company: properties });
+};
+
+
+/**
+ * Convert a date to Intercom's format.
+ *
+ * @param {Date} date
+ * @return {Number}
+ */
+
+function convertDate (date) {
+  return Math.floor(date/1000);
+}
 });
 require.register("analytics/lib/providers/keen-io.js", function(exports, require, module){
 
@@ -4678,17 +4630,19 @@ Keen.prototype.initialize = function (options, ready) {
 /**
  * Identify.
  *
+ * TODO: migrate from old `userId` to simpler `id`
+ *
  * @param {String} id (optional)
  * @param {Object} traits (optional)
  * @param {Object} options (optional)
  */
 
 Keen.prototype.identify = function (id, traits, options) {
-  var globals = {};
-  if (id) globals.userId = id;
-  if (traits) globals.traits = traits;
+  var user = {};
+  if (id) user.userId = id;
+  if (traits) user.traits = traits;
   window.Keen.setGlobalProperties(function() {
-    return { user: globals };
+    return { user: user };
   });
 };
 
@@ -6273,6 +6227,7 @@ module.exports = Provider.extend({
 
 
 
+
 require.alias("avetisk-defaults/index.js", "analytics/deps/defaults/index.js");
 require.alias("avetisk-defaults/index.js", "defaults/index.js");
 
@@ -6285,8 +6240,6 @@ require.alias("component-cookie/index.js", "cookie/index.js");
 
 require.alias("component-each/index.js", "analytics/deps/each/index.js");
 require.alias("component-each/index.js", "each/index.js");
-require.alias("component-to-function/index.js", "component-each/deps/to-function/index.js");
-
 require.alias("component-type/index.js", "component-each/deps/type/index.js");
 
 require.alias("component-event/index.js", "analytics/deps/event/index.js");
@@ -6321,8 +6274,6 @@ require.alias("ianstormtaylor-is-empty/index.js", "ianstormtaylor-is/deps/is-emp
 require.alias("ianstormtaylor-map/index.js", "analytics/deps/map/index.js");
 require.alias("ianstormtaylor-map/index.js", "map/index.js");
 require.alias("component-each/index.js", "ianstormtaylor-map/deps/each/index.js");
-require.alias("component-to-function/index.js", "component-each/deps/to-function/index.js");
-
 require.alias("component-type/index.js", "component-each/deps/type/index.js");
 
 require.alias("jkroso-equals/index.js", "analytics/deps/equals/index.js");
@@ -6345,6 +6296,13 @@ require.alias("component-type/index.js", "segmentio-bind-all/deps/type/index.js"
 require.alias("segmentio-bind-all/index.js", "segmentio-bind-all/index.js");
 require.alias("segmentio-canonical/index.js", "analytics/deps/canonical/index.js");
 require.alias("segmentio-canonical/index.js", "canonical/index.js");
+
+require.alias("segmentio-convert-dates/index.js", "analytics/deps/convert-dates/index.js");
+require.alias("segmentio-convert-dates/index.js", "convert-dates/index.js");
+require.alias("ianstormtaylor-is/index.js", "segmentio-convert-dates/deps/is/index.js");
+require.alias("component-type/index.js", "ianstormtaylor-is/deps/type/index.js");
+
+require.alias("ianstormtaylor-is-empty/index.js", "ianstormtaylor-is/deps/is-empty/index.js");
 
 require.alias("segmentio-extend/index.js", "analytics/deps/extend/index.js");
 require.alias("segmentio-extend/index.js", "extend/index.js");
@@ -6376,8 +6334,6 @@ require.alias("ianstormtaylor-is-empty/index.js", "ianstormtaylor-is/deps/is-emp
 require.alias("segmentio-on-body/index.js", "analytics/deps/on-body/index.js");
 require.alias("segmentio-on-body/index.js", "on-body/index.js");
 require.alias("component-each/index.js", "segmentio-on-body/deps/each/index.js");
-require.alias("component-to-function/index.js", "component-each/deps/to-function/index.js");
-
 require.alias("component-type/index.js", "component-each/deps/type/index.js");
 
 require.alias("segmentio-on-error/index.js", "analytics/deps/on-error/index.js");