+++ History.md
@@ -1,6 +1,7 @@
 0.11.16 - September 3, 2013
 ---------------------------
 * Woopra: fixed default pageview properties
+* Intercom: cleaned up identify logic and tests
 
 0.11.15 - September 2, 2013
 ---------------------------